backend/app/matting/service.py
from PIL import Image
import os

def process_image(input_path, output_path):
    """
    Process the image to remove background
    
    In a real implementation, this would use a sophisticated algorithm,
    possibly involving machine learning models for image segmentation.
    For this prototype, we'll implement a simple placeholder function.
    
    Args:
        input_path: Path to the original image
        output_path: Path where processed image will be saved
        
    Returns:
        bool: True if processing was successful, False otherwise
    """
    try:
        # Open the image
        img = Image.open(input_path)
        
        # TODO: Replace with actual background removal algorithm
        # For now, we'll just save the original as a placeholder
        img.save(output_path)
        
        return True
    except Exception as e:
        print(f"Error processing image: {e}")
        return False