pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=5.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "stripe>=11.6.0",
]
